update metabolomics plotting
